cp-100/requirements.txt
Flask==0.12.2
gunicorn==19.7.1

google-cloud-language==1.0.1
google-cloud-datastore==1.6.0
google-cloud-vision==0.33.0
google-cloud-storage==1.8.0
